add a build struct
