+++ vibrato/src/dictionary/builder.rs
@@ -1,21 +1,24 @@
+//! Builders for [`Dictionary`].
 use std::io::Read;
 
 use crate::dictionary::connector::{MatrixConnector, RawConnector};
 use crate::dictionary::{
-    CharProperty, ConnIdMapper, Connector, ConnectorWrapper, Dictionary, DictionaryInner, LexType,
-    Lexicon, UnkHandler,
+    CharProperty, ConnectorWrapper, Dictionary, DictionaryInner, LexType, Lexicon, UnkHandler,
 };
 use crate::errors::{Result, VibratoError};
 
 use super::lexicon::RawWordEntry;
 
-impl Dictionary {
+/// Builder for [`Dictionary`] from system lexicon entries.
+pub struct SystemDictionaryBuilder {}
+
+impl SystemDictionaryBuilder {
     pub(crate) fn new(
         system_word_entries: &[RawWordEntry],
         connector: ConnectorWrapper,
         char_prop: CharProperty,
         unk_handler: UnkHandler,
-    ) -> Result<Self> {
+    ) -> Result<Dictionary> {
         let system_lexicon = Lexicon::from_entries(system_word_entries, LexType::System)?;
 
         if !system_lexicon.verify(&connector) {
@@ -31,7 +34,7 @@ impl Dictionary {
             ));
         }
 
-        Ok(Self {
+        Ok(Dictionary {
             data: DictionaryInner {
                 system_lexicon,
                 user_lexicon: None,
@@ -44,7 +47,7 @@ impl Dictionary {
         })
     }
 
-    /// Creates a new instance from readers in the MeCab format.
+    /// Creates a new [`Dictionary`] from readers of system entries in the MeCab format.
     ///
     /// # Arguments
     ///
@@ -61,7 +64,7 @@ impl Dictionary {
         connector_rdr: C,
         char_prop_rdr: P,
         unk_handler_rdr: U,
-    ) -> Result<Self>
+    ) -> Result<Dictionary>
     where
         S: Read,
         C: Read,
@@ -83,7 +86,8 @@ impl Dictionary {
         )
     }
 
-    /// Creates a new instance from readers with the detailed bi-gram information.
+    /// Creates a new [`Dictionary`] from readers of system entries
+    /// with the detailed bi-gram information.
     ///
     /// # Arguments
     ///
@@ -104,7 +108,7 @@ impl Dictionary {
         bigram_cost_rdr: C,
         char_prop_rdr: P,
         unk_handler_rdr: U,
-    ) -> Result<Self>
+    ) -> Result<Dictionary>
     where
         S: Read,
         R: Read,
@@ -128,68 +132,6 @@ impl Dictionary {
             unk_handler,
         )
     }
-
-    /// Resets the user dictionary from a reader.
-    ///
-    /// # Arguments
-    ///
-    ///  - `user_lexicon_rdr`: A reader of a lexicon file `*.csv` in the MeCab format.
-    ///                        If `None`, clear the current user dictionary.
-    ///
-    /// # Errors
-    ///
-    /// [`VibratoError`] is returned when an input format is invalid.
-    pub fn user_lexicon_from_reader<R>(mut self, user_lexicon_rdr: Option<R>) -> Result<Self>
-    where
-        R: Read,
-    {
-        if let Some(user_lexicon_rdr) = user_lexicon_rdr {
-            let mut user_lexicon = Lexicon::from_reader(user_lexicon_rdr, LexType::User)?;
-            if let Some(mapper) = self.data.mapper.as_ref() {
-                user_lexicon.do_mapping(mapper);
-            }
-            if !user_lexicon.verify(self.connector()) {
-                return Err(VibratoError::invalid_argument(
-                    "user_lexicon_rdr",
-                    "user_lexicon_rdr includes invalid connection ids.",
-                ));
-            }
-            self.data.user_lexicon = Some(user_lexicon);
-        } else {
-            self.data.user_lexicon = None;
-        }
-        Ok(self)
-    }
-
-    /// Edits connection ids with the given mappings.
-    ///
-    /// # Arguments
-    ///
-    ///  - `lmap/rmap`: An iterator of mappings of left/right ids, where
-    ///                 the `i`-th item (1-origin) indicates a new id mapped from id `i`.
-    ///
-    /// # Errors
-    ///
-    /// [`VibratoError`] is returned when
-    ///  - a new id of [`BOS_EOS_CONNECTION_ID`](crate::common::BOS_EOS_CONNECTION_ID)
-    ///    is included,
-    ///  - new ids are duplicated, or
-    ///  - the set of new ids are not same as that of old ids.
-    pub fn mapping_from_iter<L, R>(mut self, lmap: L, rmap: R) -> Result<Self>
-    where
-        L: IntoIterator<Item = u16>,
-        R: IntoIterator<Item = u16>,
-    {
-        let mapper = ConnIdMapper::from_iter(lmap, rmap)?;
-        self.data.system_lexicon.do_mapping(&mapper);
-        if let Some(user_lexicon) = self.data.user_lexicon.as_mut() {
-            user_lexicon.do_mapping(&mapper);
-        }
-        self.data.connector.do_mapping(&mapper);
-        self.data.unk_handler.do_mapping(&mapper);
-        self.data.mapper = Some(mapper);
-        Ok(self)
-    }
 }
 
 #[cfg(test)]
@@ -203,7 +145,7 @@ mod tests {
         let char_def = "DEFAULT 0 1 0";
         let unk_def = "DEFAULT,0,0,100,*";
 
-        let result = Dictionary::from_readers(
+        let result = SystemDictionaryBuilder::from_readers(
             lexicon_csv.as_bytes(),
             matrix_def.as_bytes(),
             char_def.as_bytes(),
@@ -220,7 +162,7 @@ mod tests {
         let char_def = "DEFAULT 0 1 0";
         let unk_def = "DEFAULT,1,1,100,*";
 
-        let result = Dictionary::from_readers(
+        let result = SystemDictionaryBuilder::from_readers(
             lexicon_csv.as_bytes(),
             matrix_def.as_bytes(),
             char_def.as_bytes(),
