+++ vibrato/src/tests/tokenizer.rs
@@ -1,4 +1,4 @@
-use crate::dictionary::Dictionary;
+use crate::dictionary::SystemDictionaryBuilder;
 use crate::Tokenizer;
 
 const LEX_CSV: &str = include_str!("./resources/lex.csv");
@@ -9,7 +9,7 @@ const UNK_DEF: &str = include_str!("./resources/unk.def");
 
 #[test]
 fn test_tokenize_tokyo() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -43,7 +43,7 @@ fn test_tokenize_tokyo() {
 
 #[test]
 fn test_tokenize_kyotokyo() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -105,7 +105,7 @@ fn test_tokenize_kyotokyo() {
 
 #[test]
 fn test_tokenize_kyotokyo_with_user() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -152,7 +152,7 @@ fn test_tokenize_kyotokyo_with_user() {
 
 #[test]
 fn test_tokenize_tokyoto_with_space() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -208,7 +208,7 @@ fn test_tokenize_tokyoto_with_space() {
 
 #[test]
 fn test_tokenize_tokyoto_with_space_ignored() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -253,7 +253,7 @@ fn test_tokenize_tokyoto_with_space_ignored() {
 
 #[test]
 fn test_tokenize_tokyoto_with_spaces_ignored() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -298,7 +298,7 @@ fn test_tokenize_tokyoto_with_spaces_ignored() {
 
 #[test]
 fn test_tokenize_tokyoto_startswith_spaces_ignored() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -332,7 +332,7 @@ fn test_tokenize_tokyoto_startswith_spaces_ignored() {
 
 #[test]
 fn test_tokenize_tokyoto_endswith_spaces_ignored() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -366,7 +366,7 @@ fn test_tokenize_tokyoto_endswith_spaces_ignored() {
 
 #[test]
 fn test_tokenize_kampersanda() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -397,7 +397,7 @@ fn test_tokenize_kampersanda() {
 
 #[test]
 fn test_tokenize_kampersanda_with_user() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -430,7 +430,7 @@ fn test_tokenize_kampersanda_with_user() {
 
 #[test]
 fn test_tokenize_kampersanda_with_max_grouping() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -475,7 +475,7 @@ fn test_tokenize_kampersanda_with_max_grouping() {
 
 #[test]
 fn test_tokenize_tokyoken() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -493,7 +493,7 @@ fn test_tokenize_tokyoken() {
 /// This test is to check if the category order in char.def is preserved.
 #[test]
 fn test_tokenize_kanjinumeric() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -518,7 +518,7 @@ fn test_tokenize_kanjinumeric() {
 
 #[test]
 fn test_tokenize_empty() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
@@ -535,7 +535,7 @@ fn test_tokenize_empty() {
 
 #[test]
 fn test_tokenize_repeat() {
-    let dict = Dictionary::from_readers(
+    let dict = SystemDictionaryBuilder::from_readers(
         LEX_CSV.as_bytes(),
         MATRIX_DEF.as_bytes(),
         CHAR_DEF.as_bytes(),
