+++ vibrato/src/tokenizer.rs
@@ -284,6 +284,8 @@ impl Tokenizer {
 mod tests {
     use super::*;
 
+    use crate::dictionary::SystemDictionaryBuilder;
+
     #[test]
     fn test_tokenize_1() {
         let lexicon_csv = "自然,0,0,1,sizen
@@ -295,7 +297,7 @@ mod tests {
         let char_def = "DEFAULT 0 1 0";
         let unk_def = "DEFAULT,0,0,100,*";
 
-        let dict = Dictionary::from_readers(
+        let dict = SystemDictionaryBuilder::from_readers(
             lexicon_csv.as_bytes(),
             matrix_def.as_bytes(),
             char_def.as_bytes(),
@@ -338,7 +340,7 @@ mod tests {
         let char_def = "DEFAULT 0 1 0";
         let unk_def = "DEFAULT,0,0,100,*";
 
-        let dict = Dictionary::from_readers(
+        let dict = SystemDictionaryBuilder::from_readers(
             lexicon_csv.as_bytes(),
             matrix_def.as_bytes(),
             char_def.as_bytes(),
@@ -381,7 +383,7 @@ mod tests {
         let char_def = "DEFAULT 0 0 3";
         let unk_def = "DEFAULT,0,0,100,*";
 
-        let dict = Dictionary::from_readers(
+        let dict = SystemDictionaryBuilder::from_readers(
             lexicon_csv.as_bytes(),
             matrix_def.as_bytes(),
             char_def.as_bytes(),
@@ -424,7 +426,7 @@ mod tests {
         let char_def = "DEFAULT 0 0 3";
         let unk_def = "DEFAULT,0,0,100,*";
 
-        let dict = Dictionary::from_readers(
+        let dict = SystemDictionaryBuilder::from_readers(
             lexicon_csv.as_bytes(),
             matrix_def.as_bytes(),
             char_def.as_bytes(),
